requirements.txt
# ==================== Core Framework ====================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# ==================== LangChain & AI ====================
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
langgraph==0.0.20
llama-index==0.9.37

# ==================== Vector Databases ====================
chromadb==0.4.17
faiss-cpu==1.13.2
pinecone-client==2.2.4

# ==================== Data & Utilities ====================
pydantic==2.5.0
pydantic-settings==2.1.0
aiohttp==3.9.0
httpx==0.25.0

# ==================== Authentication & Security ====================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# ==================== Database ====================
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# ==================== Database - HANA (Optional) ====================
# hdbcli==2.19.21

# ==================== Caching (Optional) ====================
# redis==5.0.1

# ==================== Monitoring & Logging ====================
python-json-logger==2.0.7

# ==================== Development Tools ====================
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# ==================== Documentation ====================
mkdocs==1.5.3
mkdocs-material==9.4.14